requirements.txt
Pillow
pyautogui
keyboard
pygetwindow
pywin32
